Function Problem
When isSquare() function is called at line 91 within inverse function, for some reason "!other.isSquare()" is evaluating as true(its reading the inputted matrix A as not being a Square matrix even when it is 2x2).

The weird thing is, within main in the isSquare()function call at line 264, the function is working correctly and outputting that it is not square.  YET, when the inverse function is called at line 271 it's exiting due to the same matrix being evaluated as NOT being square.